Create Annotated UpSet Plot

Create an UpSet plot annotated with intersection and set sizes. Essentially a modified version of UpSet.

plot_upset(
  ...,
  mode = c("distinct", "intersect", "union"),
  top_n_comb = 10,
  scale_set_bars = FALSE,
  row_labels = character(0),
  row_names_gp = gpar(fontsize = 6.5),
  annotation_name_gp = gpar(fontsize = 6.5, fontface = "bold"),
  heatmap_args = list(),
  cell_height = unit(4, "mm"),
  cell_width = cell_height,
  extend = 0.2,
  rot = 45,
  hjust = 0,
  vjust = 0,
  bar_label_gp = gpar(fontsize = 6),
  filename = character(0),
  height = 2.5,
  width = 2.7,
  save_args = list()
)

Arguments

...

The input sets. If it is represented as a single variable, it should be a matrix/data frame or a list. If it is multiple variables, it should be name-value pairs, see Input section for explanation.

mode

The mode for forming the combination set, see Mode section.

top_n_comb

integer; number of largest intersections to display. Default is 10.

scale_set_bars

logical;

row_labels

character; labels for each set (row of UpSet plot). Defaults to rownames of make_comb_mat output.

row_names_gp

graphical parameters specified by gpar. Modifies row_labels.

annotation_name_gp

graphical parameters specified by gpar. Passed to HeatmapAnnotation. Modifies the barplot titles.

heatmap_args

list; additional arguments passed to Heatmap.

cell_height

unit object; the height of each heatmap cell. Default is unit(4, "mm").

cell_width

same as cell_height.

extend

numeric;

rot

numeric; angle in [0,360] to rotate set intersection labels.

hjust

numeric;

vjust

numeric;

bar_label_gp

graphical parameters specified by gpar. Passed to HeatmapAnnotation. Modifies the annotation labels.

filename

character; the filename used to save the heatmap. If character(0) (default), the heatmap will be displayed instead.

height

numeric; height of the file in units.

width

same as height.

save_args

list; additional arguments passed to the graphics device. See png for options.

Details

See UpSet for details and sections mentioned in Arguments.

References

Lex, A., Gehlenborg, N., Strobelt, H., Vuillemot, R., & Pfister, H. (2014). UpSet: Visualization of Intersecting Sets. IEEE transactions on visualization and computer graphics, 20(12), 1983--1992. https://doi.org/10.1109/TVCG.2014.2346248